+++ src/utils/getDataWithPct.js
@@ -1,6 +1,6 @@
 import { sum } from "d3-array";
 
-export default function useDataWithPct(data) {
+export default function getDataWithPct(data) {
   // calculate percentages for display
   const totalValue = sum(data.map(({ value }) => value));
   return data.map((record) => ({
