+++ src/utils/index.js
@@ -1,0 +1,2 @@
+export { default as formatAsPct } from "./formatAsPct";
+export { default as getDataWithPct } from "./getDataWithPct";
